Fix bug in process control in direct scheduler (#2721)
The direct scheduler uses `ps` for checking whether a command is
running. However, `ps` does not show processes without a controlling
terminal by default - for this you need to add the -x option:

  -x      When displaying processes matched by other options, include pro-
           cesses which do not have a controlling terminal.  This is the
           opposite of the -X option.  If both -X and -x are specified in
           the same command, then ps will use the one which was specified
           last.

Since this flag was not there, the engine thought calculations were finished
when they were not and so the engine started retrieving files prematurely.
Adding the flag solves this issue.