Remove the `CalcJobNode.options` shortcut property (#2736)
This can lead to confusion with the `get_option` method, which retrieves
actual option values set on the node, whereas `options` will return the
options of the process spec which created the node. By removing the
shortcut, one will have to do `self.process_class.spec_options` instead,
which is not so bad but makes it very clear that you are accessing a
property of the spec of the process.