Create transparent interface for the `Config` and `Profile` class
The `Config` and `Profile` classes are utility classes that represent
the configuration of an AiiDA instance, which is currently the
`config.json` file and the profile configurations that it contains.

To provide a clear interface for developers to set and retrieve certain
attributes, we abstract the structure in which their content is stored
in the configuration file on disk, such that they don't need to know
about the data representation. An additional advantage is that if the
data structure for storage on disk has to change, the class can map
these changes and the interface can remain the same.

The downside is that these classes can no longer be used for the
migrations but they will have to act on the raw data structure as they
would have been written to disk at the time of the migration.