Upgrade reentry dependency (#2741)
Through fixes in `reentry` the pinning of `pip<=19` can now be released as the
`reentry_scan` hook is now deprecated but is no longer needed, as the
`pyproject.toml` can be used to install `reentry` as a build requirement.
The only downside is that entry points will no longer be automatically scanned
when `pip<=10` is used as that version does not work with the toml. However,
a simple `reentry scan` will fix that.