Reinstate the test submit functionality for `CalcJob` (#2743)
The old `JobCalculation` had a method `submit_test` which would perform
a test submission, by mocking the remote working directory that would be
created in a real submission in a local sandbox folder. This method was
removed because the replacement class `CalcJobNode` no longer provided
methods to actually run anything. Instead this is now the responsability
of the `CalcJob` process class.

To provide similar functionality a new metadata input `dry_run` is
introduced for the `CalcJob` class. When passed in the inputs when the
process is ran, the same procedure as a normal submission is started but
the `upload_calculation` step will the local transport instead of the
targeted remote computer and will write the files to a local sandbox
folder instead of a remote working directory. After the upload task is
completed, the process is stopped instead of going to the submission
step.

This new approach has the advantage over the old implementation that the
pathway that is followed in the test submission is alsmost identical to
the code that would be executed in an actual submission. Additionally,
by using a metadata option, instead of a special launcher function or a
specific method on the `ProcessBuilder` ensures that the user does not
have to import and call a separate function, nor does the method on the
builder have to block a name in the input namespace.