Hide `verdi daemon start-circus` from auto-completion and help
This command should not be called directly by users but will rather be
invoked by `verdi daemon start`. This is accomplished by marking the
command as `hidden=True` which was implemented in `click==7.0`. However,
the attribute was being ignored by the `click-completion` package, which
was fixed in `click-completion==0.5.1` which is now added as dependency.