Documentation: add section on process specification (#2751)
This is a very general explanation of the concept of a process
specification and the port and port namespaces that populate its
`inputs` and `outputs` attributes. This is intentionally kept a bit
abstract because these concepts apply to all processes, which allows it
to be defined and explained in one place. The sections on concrete
processes can then be explained on a more generic level and simply refer
to this section for the details.