Print formatted spec for `verdi plugin list <entry_point>` (#2757)
When a valid entry point is passed an the corresponding class is a sub
class of `Process`, the `get_description` message will return the
process spec which is then passed to a function that will print it in a
human-readable formatted way to screen. It will give an overview of the
inputs, outputs and exit codes defined on the spec.