Restore non-interactive functionality of `verdi quicksetup`
The presence of the `non-interactive` option not only triggers the
command to automatically try to guess the connection method to Postgres
but it will also influence the other interactive options to simply
return the default if it was not explicitly passed. In interactive mode
the options will always trigger the prompt, even if there is a default.