New tests for export file migrations - now in separate repo "aiida-export-migration-tests".
Tests for the following files in `cmdline.utils.migration`:
- `__init__.py` in `test_migration.py`
- `v01_to_v02.py` in `test_v01_to_v02.py`
- `v02_to_v03.py` in `test_v02_to_v03.py`
- `v03_to_v04.py` in `test_v03_to_v04.py`

Found and fixed bug for migration of TrajectoryData in Rev. 1.0.26 and 1.0.27 for v0.3 to v0.4:
"symbols" was only updated for "node_attributes" not "node_attributes_conversion".

Added more representative export files for the different export versions.
They have been created using the same workflow (by @yakutovicha) in specialized QuantumMobile VMs.
AiiDA version used for export versions:
Export: 0.2, AiiDA: 0.9.1
Export: 0.3, AiiDA: 0.12.3
Export: 0.4, AiiDA: 1.0.0b2

Export version 0.1 was set to AiiDA version 0.6.0,
but AiiDA was never released on PyPi until version 0.8.0,
and the changes from export version 0.1 to 0.2 are very minor.
So a more representative export file for export version 0.1 is left out.
The export files are in the separate repo.

Added dependency for separate repo if "testing" is chosen when installing via pip.

To be pythonic and have more transparent code,
the migration steps were divided into separate functions.
This also results in the fact that a lot of migrations can now be skipped,
if no Nodes are present in the export file.