Implement workaround to fix docs build failing on python 3
The documentation builds started failing after the mechanism for loading
profiles and the database backend was changed significantly. Since then
Sphinx started failing with an exception that the attributes `run` and
`runctx` are not defined in the `profile` module. What it should be
targeting is the `cProfile` method, but somehow it ends up targeting the
`aiida.backends.profile` module. Simply adding the attributes there
fixes the problem, even though we have absolutely no idea why. Given
that the module is deprecated anyway and will soon be removed at which
point the problem should no longer exist.