Clean the implementation of `verdi quicksetup`
The logic for automatically determining the required options if not
specified, as well as reprompting if incorrect values are selected are
now implemented directly in `click` options instead of the function
body. The various database connection options are also restored for
`verdi quicksetup` that were removed in a recent commit. This allows to
override the defaults while still profiting of the automatic resource
creation.

To avoid code duplication, the reusable options are defined for use in
`verdi quicksetup`, which should not prompt for the database connection
options, and these are then extended to become fully interactive options
for `verdi setup` which will make them prompt even if there is a
default.