Implement and use a `AiidaPostgresTestCase`
With the added functionality of resetting a loaded profile and thereby
loading a different one, with potentially a different backend, allows
implementing tests for example `verdi setup` who rely on this
functionality to be implemented as a normal unit test case, instead of
an isolated and separated CI test that does not pre-load the test
profile. This allows to move the tests of `.ci/test_quicksetup` to the
`aiida.backends.tests.cmdline.commands.setup` commands.