Remove `aiida.tests` entry point group and `parser_tests`
The `aiida_tests` entry point was added as a temporary replacement until
the plugin test fixtures came available. These have been available for a
while now and since no plugins rely on the entry point group we remove
it. Additionally, the `test_parsers` was a first implementation to test
parser plugins, but this can be done more elegantly with `pytest` and
the plugin fixtures, therefore it is removed.