Prevent additional incoming links for a stored `ProcessNode` (#2777)
Process nodes are in principle only generated by the engine as soon as a
process is launched. At that time, the engine will take care of adding
incoming links for all the inputs that were passed as well as an
optional call link. After that the node is stored, but until it is
terminated and sealed, it was still possible to add additional incoming
links. This is now explicitly forbidden as it does not make sense and so
if attempted will raise a `ValueError`. As a result, many tests had to
be adapted that were violating this rule. Often the fix was just to move
the `store` call to after all links were put in place.