Change the way hash is calculated for float
Hash the float by converting it to string and only use the first 14 significant
digits. This is consistent with the way rounding error may occur is lost when
saving a float to the database.
This fix the issue where the hash of a node containing float
attributes may change when it is stored.