Deal with stale daemon PID files during stop attempts
In some circumstances, the circus PID file which stores the PID of the
running circus process, can be corrupted or not deleted properly,
causing calls to `is_daemon_running` to return a false positive.

This situation can arise if a system is shut down suddenly and so
the process is killed but the PID file is not deleted in time.
Alternatively, another process or the user may have meddled with the
PID file in some way, corrupting it.

Here, we implement a function that checks for stale PID files by
checking whether the PID contained in the PID file matches a valid
running `verdi `process. If it does not, a warning is emitted and the
PID file is deleted. This function is called when `verdi daemon stop` is
called, making it function as a hard reset in a situation of a stale PID.

Co-authored-by: Giovanni Pizzi <giovanni.pizzi@epfl.ch>