Properly unload database environment when swapping profiles
The experimental functionality to switch profile within the same
interpreter that was recently introduced, suffered a fatal flaw. Even
though the `load_profile` function correctly reset the `Manager`, which
would hold references configured for the old profile, it did not
consider database specific resources. The `Backend` instance was
properly reset, but both the Django and SqlAlchemy implementations also
configure an engine or connection to the database, based on the database
connection details specified for the profile.

These resources were not properly being reset and recreated after
changing profiles, causing the wrong database to be used. This was
discovered in the unittests for `verdi quicksetup/setup`. These tests
would create a complete temporary AiiDA instance complete with postgres
cluster and create a new profile, to which it would then switch for the
duration of the test. After the test, the original test session
variables would be restored. However, the database environment was never
actually being swapped, causing the tests not to use the temporary
database, but that of the general test session.

This problem is addressed, by implementing an `unload_dbenv` function
that destroys the database resources that are initialized by the
`load_dbenv` counterpart. This function is called when the manager is
reset, which in turn is called in `load_profile` when the profile is
switched.

For SqlAlchemy, the unloading is accomplished by disposing the current
engine and resetting the scoped session class by calling`reset_session`
which will reset the global variables in `aiida.backends.sqlalchemy`.

The `unload_dbenv` for Django has not yet been implemented. The
mechanism for loading the database environment for Django is to import
the settings file, which will define global variables based on the
current profile, which will be used by Django to create a database
connection. It is currently not clear, how, once these variables are
locked in, can be clearly reset to those based on the new profile.