Forward `submit` to `run` for `dry_run=True`
Typically one will run `dry_run=True` in combination with
`store_provenance=False`, however, that is required for `submit` to work
as it needs to keep a record in the database to communicate it to other
workers. Instead of forbidding `submit` to be used with a dry run, we
simply forward internally to `run`. This way the user does not have to
switch the launcher between testing with dry runs and actual submissions.