Require `store_provenance=True` in `submit` or raise (#2828)
The current design of submitting processes, requires them to be stored
in the database as the state is stored as a checkpoint on the node. This
means that allowing `store_provenance=False` is non-sensical and will
lead to unexpected and cryptic errors. Instead we check for this value
and raise when set to False in combination with `submit`.