Fix bug in Django migration `0023_calc_job_option_attribute_keys`
The migration changed the keys of certain node attributes but failed to
properly join the matched attributes on the matched nodes. This could
cause attributes with the target keys to be affected even if the
corresponding node would not have matched the filter. The chances of
this having occurred are minimal to non-existent, as it would require
non calculation job nodes with these particular keys to have existed, so
we simply fix the migration and leave it at that.