Fix bug in SqlAlchemy migration `7ca08c391c49_calc_job_option_attribute_keys`
The migration incorrectly converted the dictionary attributes with the
new keys `resources` and `environment_variables` to text, because the
wrong operator `->>` was used. This operator gets the JSONB value as
text, whereas `->` keeps the type, which is the one that should have
been used. This was missed by the tests, because the tests undid the
type conversion using `ast`. Collossal fail...