Update dependencies `Jinja2` and `SqlAlchemy` for vulnerabilities
When upgrading `SqlAlchemy` the migration tests started failing because
the migration that drops the legacy workflow tables in the downgrade
reinstates a uniqueness constraint that exceeds the 63 maximum character
limit. Since these tables are dropped anyway and only created in the
downgrade for migration testing purposes, we should be safe in changing
the name of the identifier.