Clean work directory in Jenkins (#2855)
In Jenkins the work directory was not being cleaned.
Besides keeping, in the long term, GBs of old files,
this was problematic for tests because the work
directories are often reused, and this is a problem
e.g. if a folder is deleted on git but .pyc files
are left in.

This commit both runs a clean command from git
before running (so it removes any file that is not
under git control) and also cleans the workspace
directory at the very end of the job to free up space.