Fix bug in creation of `SinglefileData` in `retrieve_calculation` (#2860)
This functionality in the `retrieve_calculation` function of the
calculation job execmanager automatically creates `SinglefileData` nodes
based on the `retrieve_singlefile_list` set on the `CalcJobNode`. The
constructor requires the `file` argument which was not passed. The
constructor will automatically place the file in the repository so the
following call to `put_object_from_file` is superfluous.