Allow global submit from within work functions (#2869)
Recently, a limitation was placed on the global `submit` launcher
function and an exception would be raised if it was called from within
the scope of another process, because for this the `self.submit` method
of the process should be used instead. Note that this only applies to
workflow-like processes, since calculation-like processes *cannot* not
even call other processes. So this applies to work functions and work
chains. However, the process instance is not available within the body
of a work function implementation, so one *cannot* do `self.submit` and
in fact the global submitter should be used. Here we update the guard
to exclude work functions from this limitation.