Automatically set `CalcInfo.uuid` in `CalcJob.run`
The `uuid` field in the `CalcInfo` data structure which is built and
returned by the `CalcJob.prepare_for_submission` implementation, is used
subsequently by the engine to create the remote working directory.
Previously, the implementer of `prepare_for_submission` was required to
set it, but this always boiled down to setting it to the UUID of the
process node. Since this is the only behavior that should happen anyway,
we might as well have the engine take care of this, making the
implementation of the `prepare_for_submission` for plugins simpler.