Include only input links in creating a process restart builder
The `Process.get_builder_restart` was populating the builder by getting
all the nodes with incoming links, including call links which of course
are not defined in the spec and so would cause an exception. Simply by
adding a link filter to only get input data nodes, this problem is fixed.