Fix false negative of postgres connection in `verdi status` (#2924)
The output of `verdi status` was reporting a problem in connecting to
the postgres database even though all other commands seemed to work. The
problem was with the `Postgres.from_profile` used to perform the check.
This utility does on purpose not include the database name, username and
password, because it is supposed to use the systems superuser instead.
To check whether the current profile can access the database using the
connection details specified within it, we simply try to get a cursor
from the backend, which will except if it cannot connect to the database.