Ensure `mtime` update is triggered for Django
The `ModelWrapper` utility class ensures that a database model instance
is automatically saved if a model field is updated and the instance is
already stored. This prevents the implementation from having to call
`save` manually everytime a field is updated. The Django implementation
uses the `update_fields` keyword of the `save` method, to update only
the field affected. However, this circumvents the `auto_now` of the
`mtime` field. This bug is currently not visible, because in the model
wrapper currently also always updates the node version after a model
field is updated of a stored instance. This, however, uses a normal
`save` and so does properly update the `mtime`. Once this updating of
the node version would have been removed, this bug will be revealed.