Acknowledge process task if corresponding node cannot be loaded (#2936)
Before this commit the `ProcessLauncher._continue` method was raising a
`plumpy.TaskRejected` exception if the node of a process task cannot be
loaded, either because it does not exist or it cannot be uniquely
resolved. However, this will cause RabbitMQ to requeue the task, causing
the task to be resent. This will result in the task ping-ponging between
RabbitMQ and the daemon workers without end. This situation typically
occurs when a user deletes a process node while it has not yet properly
terminated and therefore the tasks has not yet been acknowledged. This
situation is unrecoverable so the task should simply be acknowledged.