Simplify the `User` database model by dropping unused columns
The following columns are dropped:

 * `password`
 * `date_joined`
 * `groups`
 * `is_active`
 * `is_staff`
 * `is_superuser`
 * `last_login`
 * `user_permissions`

These columns were originally added because the base user class of the
Django framework was used, which provided the ORM layer. However, these
attributes are not used at all and Django does not require them in order
to work properly. Therefore, we remove them from the models for both
backends.

Note that in principle, the `contrib` and `auth` applications currently
installed for Django, could also be removed without affecting the
functionality of `aiida-core` since it no longer requires the built in
authentication application of Django. However, since the initial
migration schema still references it, we cannot yet remove it fully.
If we were to reset the database migration schema, we could get rid of
these unnecessary dependencies.