Drop column `transport_params` from the `Computer` database model
The column was intended to contain customized parameters to be used to
configure the transport of a computer. However, these fit better in the
`AuthInfo`, as these settings can change per user. The design is to keep
the `Computer` as generic as possible as to not be user dependent, such
that its configuration can be easily shared and reused. Generic computer
configuration can go in the `metadata` column. User specific settings
need to go in the `AuthInfo.metadata` instead.