Update dependency `kiwipy==0.5.1` with critical bug patch
The persistence of a task message, which represents a process that is to
be executed, is guaranteed by RabbitMQ, but only if the exchange, queue
and the task message itself are configured correctly. Up till now the
exchange and task queue were properly set to be durable, however, the
individual messages were sent with the default delivery mode, which
means non-persistent. This caused tasks to be lost on broker restart.
The patch release 0.5.1 of `kiwipy` fixes this by now always marking
task messages as persistent.