Improve `safe_interval` validation `verdi computer configure`
The computer configuration command for the transport specified that
an integer was required. However, a float value makes more sense. In
addition to this change, an additional validation callback was added to
make sure that the interval is positive.

The addition of the validation through the default option callback
mechanism required a small adaptation to the `InteractiveOption`.
The behavior of error handling was different if the option was
explicitly defined on the command line but without an explicit
`--non-interactive` flag. If a string value was passed, the type
validation would fail and prompt loop was never entered. However, if a
negative number was passed, the type validation would pass, but then the
normal callback would fail and in this case the prompt would start.
However, this was with the context improperly configured and so
exceptions would occur in the contextual defaults.