Add possibility to pass `pgtest` configuration to text fixture manager (#2941)
The test fixture manager takes care of setting up a complete Postgres cluster if
a pre-configured test profile is not available. However, the parameters for this
`pgtest` utility were not configurable. This was problematic for systems that
needed to use for example another `pg_ctl` path than the default. Here we adapt
the fixture manager to be able to pass this configuration through.