Fix bug in the Django attributes and extras to JSONB migration
The migrations of the old attributes and extras from the EAV tables to a
JSONB column on the node table, used a lazy fetch method to fetch nodes
in batches. However, the consecutive fetches do not sort on anything so
the same nodes may be fetched multiple times and as a result, some nodes
are never fetched. Their attributes and extras are therefore then also
never migrated. This caused certain nodes to have lost their data after
the migration. Sorting the lazy fetches by the pk of the nodes,
guarantees that each and every node is migrated, and only once.