Gracefully exit `verdi export inspect` when archive is corrupt
A corrupt archive is any export archive that does not contain all the
required files or any of its files have a corrupt format. The `Archive`
utility class is improved to detect problems with archives and will now
throw a `CorruptArchive` exception which `verdi export inspect` will now
catch and exit gracefully with a critical error message.

The `with_dbenv` decorator was also removed from the command as it does
not need the database backend, speeding up the command.