Ensure that stacktrace is included in transport task exceptions
When a transport task of a calculation job excepts, the expontential
back off retry wrapper calls the logger of the process node with the
exception. This will trigger the `DbLogHandler` such that the log
message is entered into the database with a link to the process node.
However, only the message was included and not the stack trace which
made it difficult to determine the cause from `verdi process report`.
To fix this, in the `Log.objects.create_from_record` we attempt to
format a traceback from the `exc_info` if it is included in the record
which will be the case for an exception log.

This commit also fixes a bug in the `ProcessLauncher._continue` method
that was calling the `create_future` method of the communicator, which
has long been removed. Instead a normal future should just be created.