Fix problem in validation of `InteractiveOption` CLI option type
The option allows to specify a callback function to perform validation
in addition to the type. However, in the prompt loop, only the normal
type validation was called. This caused the validation in the callback
to be skipped if called more than once, allowing illegal values to be
accepted in interactive mode if simply passed more than once.

Additionally, the `non_interactive_default` setting for the
`safe_interval` option of `verdi computer configure` in combination with
the implementation of the callback function, which ignored `None`
values, allowed `None` values to be set in non-interactive mode. The
callback now properly checks for the correct type and raises an error if
`None` is passed.