Add `errors_handled` updatable attribute to `WorkChainNode`
This new updatable attribute for the `WorkChainNode` class will be
useful once the `BaseRestartWorkChain` and `register_error_handler`
concepts will be added to `aiida-core`. These utilities provide a basic
framework to quickly write powerful work chains around `CalcJob`
implementations. Through the `register_error_handler` methods can be
attached to the `BaseRestartWorkChain` implementation to extend the
error handling capability in the case of a failed calculation. To make
it possible to query for work chains that had certain error handlers
triggered during their execution, their method names need to be added to
some list in the attributes. This is the intended purpose for the new
`errors_handled` attribute.