Correct decorator implementation using the `wrapt` library (#2991)
The naive implementation of a decorator will break some of the
functionality of the functions they wrap. For example, the original
docstring will be lost. The module `functools` provides the `wraps`
decorator to fix this problem, but this is not the only issue. In
addition to the docstring, other properties, such as the function
signature are also lost. This means that using the `inspect` module on
decorated functions is hopelessly broken. The `wrapt` library provides
the `decorator` decorator that simplifies writing decorators, while
ensuring that all the important properties of wrapped function are
maintained.