Set the `process_status` if `WorkChain` goes to `Wait` for subprocess (#3000)
As soon as an `awaitable` is added to a `WorkChain` instance, its
process status is updated. This process status will contain a human
readable string with the pks of the sub processes for which it is
waiting. If an `awaitable` is removed and no awaitables remain, the
process status of the work chain node is cleared.