Remove all support for datetime objects in attributes/extras
Datetime objects were already not supported for SqlAlchemy since that
always used JSONB, which does not have native support for this datatype.
But now that Django also is migrated to use JSONB, the old support of
the EAV implementation for datetime objects needs to be removed.

The `clean_value` utility function that pre-processes each value that is
entered to be stored in the database now checks a whitelist of
acceptable value, which are essentially json-serializable, or raises an
exception for anything else that is not recognized.

The `aiida.common.timezone` module provides two utility functions named
`datetime_to_isoformat` and `isoformat_to_datetime` to serialize and
deserialize datetime objects to strings, respectively, that can be
stored in the database.

Co-authored-by: Giovanni Pizzi <giovanni.pizzi@epfl.ch>