Remove stored restriction on settings node extras
Extras could originally only be set and mutated on stored nodes. The
reason for this limitation was that the original implementation of
extras in Django used the custom EAV schema to set arbitrary key value
pairs on a node. Since this requires a foreign key of the node for any
extra to be set, the corresponding node needed to be stored. Attributes
that used the same mechanism had this restriction alleviated by using an
in memory cache on the node instance while it was not stored. Since this
cache was never implemented for extras it was simply not allowed to set
them on unstored nodes.

Now that the Django implementation has switched from the custom EAV
schema for attributes and extras to a JSONB field, the foreign key
restriction no longer exists and the database model instance naturally
comes with a cache while it is not stored. Therefore the restriction on
nodes being stored for extras being mutable can now be lifted.

Co-authored-by: Giovanni Pizzi <giovanni.pizzi@epfl.ch>