Reimplement the `Node` and `BackendNode` attributes extras interface
Now that the Django implementation for the node attributes and extras
uses a JSONB field, just like SqlAlchemy, the interface for interacting
with these node properties can be homogenized and corrected.

The clear division between the front and back end classes `Node` and
`BackendNode` allows us to define a clear separation of responsabilities
as well.

 * `BackendNode`: ensure serializability of values stored in the fields
   as they will have to be stored as JSON. The implementation needs to
   ensure that cleaning of values is done as little as possible and
   scales linearly when appending to existing values.

 * `Node`: implements AiiDA specific business logic such as the rules
   for the mutability of attributes as well as the validity of key names.

The use of the `BackendNode` allows us to remove all business logic from
the database models, which in turn allows us to write exhaustive tests
for the interface in a backend independent way.

Co-authored-by: Giovanni Pizzi <giovanni.pizzi@epfl.ch>