Remove superfluous `from_dbmodel` override from backend implementations
The Django and SqlAlchemy implementations of the `BackendUser` and
`BackendComment` classes reimplemented the `from_dbmodel` method.
However, the implementation was exactly that of the base class
`BackendEntity` so it is better to rely on that. All that had to be done
was to set the `ENTITY_CLASS` class attribute to the correct class.