Add data migration for legacy `JobCalcState` attributes
The old `JobCalculation` maintained its state in the `DbCalcState` table
whose value was also stored as an attribute with the key `state` as a
proxy. With the change to processes, all `JobCalculation` nodes have
been converted to `CalcJobNodes` but did not receive any of the typical
process attributes such as `process_state` and `exit_status`. Moreover,
they kept their `state` attribute which now contains deprecated values.
The `state` attribute is still set for `CalcJobs` but it serves a
different purpose as it is now a sub state to the process state, while
the process is active for extra granularity.

Here we add a data migration for those old calculation jobs where some
process attributes are inferred from the old state attribute, which is
then discarded. This will make sure that even for old migration
calculations, commands like `verdi process list` will show some
information that allows a user to discern what the entry was.